requirements.txt
# --- UI & Excel Exports (Front-End) ---
streamlit
openpyxl

# --- AI & PDF Processing (Extractor) ---
openai
python-dotenv
PyMuPDF

# --- Web Scraping & Data Eng (Existing Tools) ---
beautifulsoup4
bs4
certifi
cffi
charset-normalizer
cryptography
curl_cffi
frozendict
idna
multitasking
numpy
pandas
pdfminer.six
pdfplumber
peewee
pillow
platformdirs
protobuf
pycparser
pypdfium2
pyrate-limiter
python-dateutil
pytz
requests
sec-edgar-downloader
six
soupsieve
typing_extensions
urllib3
websockets
yfinance